Fix /docs symlink issue
The GitHub pages website, which is served from the `docs` directory
wasn't building, because the JS files are actually just symlinks out to
the various build directories that `tsc` outputs to.

Github Pages doesn't support symlinks to files outside of `docs`.

This gets around the issue by storing the website with its symlinked
javascript files in a new directory, `website`. We then build the GitHub
Pages site with `cp -r website docs`. Copying copies the symlinked
files, removing the symlinks.

The website in `website` is currently just static HTML and JS files, but
eventually it could be a Hugo or Jekyll site