Add "White" to the list of colors
"White" is useful to not draw a dot (assuming the background is white as well).

This feature is e.g. used in the ["You Killed a Bear"](https://craigmbooth.com/projects/ykab/) featured game linked in the README, but currently this floods the console with error messages ("no CSS color defined for WHITE").

This PR does not update the relevant documentation.